+++ R/MaximizeParsimony.R
@@ -166,8 +166,8 @@
 #' 
 #' @template MRS
 #' 
-#' @importFrom cli cli_alert cli_alert_danger cli_alert_info cli_alert_success cli_alert_warning
-#' cli_h1 
+#' @importFrom cli cli_alert cli_alert_danger cli_alert_info cli_alert_success
+#' cli_alert_warning cli_h1 
 #' cli_progress_bar cli_progress_done cli_progress_update
 #' @importFrom fastmatch fmatch
 #' @importFrom stats runif
@@ -670,7 +670,7 @@ MaximizeParsimony <- function (dataset, tree,
              "TBR depth {ceiling(searchIter)}; ",
              "max. {ceiling(searchHits)} hits; ",
              "k = {concavity}.")
-    cli_alert("{ .DateTime()}: Score to beat: {.strong {signif(bestScore)}}")
+    .Info(1L, "{ .DateTime()}: Score to beat: {.strong {signif(bestScore)}}")
     
     iter <- 0L
     while (iter < ratchIter) {